mcp-servers/tools-python/requirements.txt
mcp>=1.0.0
greynoise>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.0

